fix redo undo issue
what
-----
when redo/undo action, the deleted instance will be resumed, but the instance could not be removed by undo/redo.

why
-----
since the redo/undo to work for create/delete and unable to do the reserve operation.
(This might be a arch issue, will fix it by the next big update)

how
-----
To forbid the redo/undo on deleted and craeted instance. Current operation only for existing state.

branch: dev
user: victor.yao