fix issue when redo/undo is not support
what
----
when redo/undo is not supported, the subscribe and listener will not work.

why
----
currently, the changed values are gotten from redo/undo stack, if the redo/undo is not enabled, the changes different will be empty that means no change.

how
----
to add a return value from applying changes and not get the result from redo/undo stack.

branch: dev
user: senyun.yao