support template actions and selectors & version => 1.0.0-beta.0
since for actual use case, we need to support a unknown store type actions and selectors. So, use a template to auto adapt the action/selector implementation.

Additional: from the testing of tianyu-shell, update the tianyu-store version tobe ready for release version => 1.0.0, and also in beta with starting 0.

branch: dev
user: victor.yao